Rework SSL/TLS documentation.
Motivation:

The current SSL/TLS documentation is coupled to the Vert.x Net documentation.

Most of the SSL/TLS configuration has been encapsulated in SSLOptions/ClientSSLOptions/ServerSSLOptions.

We should provide a base documentation for configuring those options and keep the Vert.x Net configuration to a minimum and refer to the documentation of those options.

Changes:

Extracted the Vert.x Net client/server SSL configuration part that is generic and not coupled to this section of the documentation in a new top level section that generically documents client/server SSL options.

Keep in Vert.x Net documentation the specific documentation.

Added examples for ConnectOptions SSL configuration.

Move the TLS upgrade documentation to the Net SSL documentation section.
