+++ vertx-core/src/main/asciidoc/server_ssl.adoc
@@ -1,0 +1,163 @@
+=== Server SSL/TLS configuration
+
+==== Specifying key/certificate for the server
+
+SSL/TLS servers usually provide certificates to clients in order to verify their identity to clients.
+
+Certificates/keys can be configured for servers in several ways:
+
+The first method is by specifying the location of a Java key-store which contains the certificate and private key.
+
+Java key stores can be managed with the http://docs.oracle.com/javase/6/docs/technotes/tools/solaris/keytool.html[keytool]
+utility which ships with the JDK.
+
+The password for the key store should also be provided:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example17}
+----
+
+Alternatively you can read the key store yourself as a buffer and provide that directly:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example18}
+----
+
+Key/certificate in PKCS#12 format (http://en.wikipedia.org/wiki/PKCS_12), usually with the `.pfx`  or the `.p12`
+extension can also be loaded in a similar fashion than JKS key stores:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example19}
+----
+
+Buffer configuration is also supported:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example20}
+----
+
+Another way of providing server private key and certificate separately using `.pem` files.
+
+[source,$lang]
+----
+{@link examples.SslExamples#example21}
+----
+
+Buffer configuration is also supported:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example22}
+----
+
+Vert.x supports reading of unencrypted RSA and/or ECC based private keys from PKCS8 PEM files.
+RSA based private keys can also be read from PKCS1 PEM files.
+X.509 certificates can be read from PEM files containing a textual encoding of the certificate as defined by
+https://tools.ietf.org/html/rfc7468#section-5[RFC 7468, Section 5].
+
+WARNING: Keep in mind that the keys contained in an unencrypted PKCS8 or a PKCS1 PEM file can be extracted by
+anybody who can read the file. Thus, make sure to put proper access restrictions on such PEM files in order to
+prevent misuse.
+
+Finally, you can also load generic Java keystore, it is useful for using other KeyStore implementations
+like Bouncy Castle:
+
+[source,$lang]
+----
+{@link examples.SslExamples#exampleBKS}
+----
+
+==== Specifying trust for the server
+
+SSL/TLS servers can use a certificate authority in order to verify the identity of the clients.
+
+Certificate authorities can be configured for servers in several ways:
+
+Java trust stores can be managed with the http://docs.oracle.com/javase/6/docs/technotes/tools/solaris/keytool.html[keytool]
+utility which ships with the JDK.
+
+The password for the trust store should also be provided:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example23}
+----
+
+Alternatively you can read the trust store yourself as a buffer and provide that directly:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example24}
+----
+
+Certificate authority in PKCS#12 format (http://en.wikipedia.org/wiki/PKCS_12), usually with the `.pfx`  or the `.p12`
+extension can also be loaded in a similar fashion than JKS trust stores:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example25}
+----
+
+Buffer configuration is also supported:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example26}
+----
+
+Another way of providing server certificate authority using a list `.pem` files.
+
+[source,$lang]
+----
+{@link examples.SslExamples#example27}
+----
+
+Buffer configuration is also supported:
+
+[source,$lang]
+----
+{@link examples.SslExamples#example28}
+----
+
+==== Server Name Indication (SNI)
+
+Server Name Indication (SNI) is a TLS extension by which a client specifies a hostname attempting to connect: during
+the TLS handshake the client gives a server name and the server can use it to respond with a specific certificate
+for this server name instead of the default deployed certificate.
+
+If the server requires client authentication the server can use a specific trusted CA certificate depending on the
+indicated server name.
+
+When SNI is active the server uses
+
+* the certificate CN or SAN DNS (Subject Alternative Name with DNS) to do an exact match, e.g `www.example.com`
+* the certificate CN or SAN DNS certificate to match a wildcard name, e.g `*.example.com`
+* otherwise the first certificate when the client does not present a server name or the presented server name cannot be matched
+
+When the server additionally requires client authentication:
+
+* when {@link io.vertx.core.net.JksOptions} is {@link io.vertx.core.net.NetServerOptions#setTrustOptions set on trust options} then an exact match with the trust store alias is done
+* otherwise the available CA certificates are used in the same way as if no SNI is in place
+
+CAUTION: the server fails to bind with an error reporting an incorrect alias
+
+You can enable SNI on the server by setting {@link io.vertx.core.net.NetServerOptions#setSni(boolean)} to `true` and
+configured the server with multiple key/certificate pairs.
+
+Java KeyStore files or PKCS12 files can store multiple key/cert pairs out of the box.
+
+[source,$lang]
+----
+{@link examples.SslExamples#configureSNIServer}
+----
+
+{@link io.vertx.core.net.PemKeyCertOptions} can be configured to hold multiple entries:
+
+[source,$lang]
+----
+{@link examples.SslExamples#configureSNIServerWithPems}
+----
